+++ src/main/java/nc/recipe/multiblock/FissionIrradiatorRecipes.java
@@ -16,19 +16,19 @@ public class FissionIrradiatorRecipes extends BasicRecipeHandler {
 	
 	@Override
 	public void addRecipes() {
-		addRecipe(Lists.newArrayList("ingotThorium", "dustThorium"), "dustTBP", 160000, fission_irradiator_heat_per_flux[0], fission_irradiator_efficiency[0], 0, -1, RadSources.THORIUM);
-		addRecipe(Lists.newArrayList("ingotTBP", "dustTBP"), "dustProtactinium233", 2720000, fission_irradiator_heat_per_flux[1], fission_irradiator_efficiency[1], 0, -1, RadSources.TBP);
-		addRecipe(Lists.newArrayList("ingotBismuth", "dustBismuth"), "dustPolonium", 1920000, fission_irradiator_heat_per_flux[2], fission_irradiator_efficiency[2], 0, -1, RadSources.BISMUTH);
+		addRecipe(Lists.newArrayList("ingotThorium", "dustThorium"), "dustTBP", 160000L, fission_irradiator_heat_per_flux[0], fission_irradiator_efficiency[0], 0L, -1L, RadSources.THORIUM);
+		addRecipe(Lists.newArrayList("ingotTBP", "dustTBP"), "dustProtactinium233", 2720000L, fission_irradiator_heat_per_flux[1], fission_irradiator_efficiency[1], 0L, -1L, RadSources.TBP);
+		addRecipe(Lists.newArrayList("ingotBismuth", "dustBismuth"), "dustPolonium", 1920000L, fission_irradiator_heat_per_flux[2], fission_irradiator_efficiency[2], 0L, -1L, RadSources.BISMUTH);
 	}
 	
 	@Override
 	protected List<Object> fixedExtras(List<Object> extras) {
 		ExtrasFixer fixer = new ExtrasFixer(extras);
-		fixer.add(Integer.class, 1);
+		fixer.add(Long.class, 1L);
 		fixer.add(Double.class, 0D);
 		fixer.add(Double.class, 0D);
-		fixer.add(Integer.class, 0);
-		fixer.add(Integer.class, -1);
+		fixer.add(Long.class, 0L);
+		fixer.add(Long.class, -1L);
 		fixer.add(Double.class, 0D);
 		return fixer.fixed;
 	}
